+++ gateway/errors_test.go
@@ -87,4 +87,38 @@ func TestWebError(t *testing.T) {
 		webError(w, r, config, err, http.StatusInternalServerError)
 		require.Equal(t, http.StatusTeapot, w.Result().StatusCode)
 	})
+
+	t.Run("Error is sent as HTML when 'Accept' header contains 'text/html'", func(t *testing.T) {
+		t.Parallel()
+
+		w := httptest.NewRecorder()
+		r := httptest.NewRequest(http.MethodGet, "/blah", nil)
+		r.Header.Set("Accept", "something/else, text/html")
+		webError(w, r, config, NewErrorStatusCodeFromStatus(http.StatusTeapot), http.StatusInternalServerError)
+		require.Equal(t, http.StatusTeapot, w.Result().StatusCode)
+		require.Contains(t, w.Result().Header.Get("Content-Type"), "text/html")
+	})
+
+	t.Run("Error is sent as plain text when 'Accept' header does not contain 'text/html'", func(t *testing.T) {
+		t.Parallel()
+
+		w := httptest.NewRecorder()
+		r := httptest.NewRequest(http.MethodGet, "/blah", nil)
+		r.Header.Set("Accept", "application/json")
+		webError(w, r, config, NewErrorStatusCodeFromStatus(http.StatusTeapot), http.StatusInternalServerError)
+		require.Equal(t, http.StatusTeapot, w.Result().StatusCode)
+		require.Contains(t, w.Result().Header.Get("Content-Type"), "text/plain")
+	})
+
+	t.Run("Error is sent as plain text when 'Accept' header contains 'text/html' and config.DisableHTMLErrors is true", func(t *testing.T) {
+		t.Parallel()
+
+		config := &Config{Headers: map[string][]string{}, DisableHTMLErrors: true}
+		w := httptest.NewRecorder()
+		r := httptest.NewRequest(http.MethodGet, "/blah", nil)
+		r.Header.Set("Accept", "something/else, text/html")
+		webError(w, r, config, NewErrorStatusCodeFromStatus(http.StatusTeapot), http.StatusInternalServerError)
+		require.Equal(t, http.StatusTeapot, w.Result().StatusCode)
+		require.Contains(t, w.Result().Header.Get("Content-Type"), "text/plain")
+	})
 }